+++ lib/util/web/non_web.dart
@@ -42,11 +42,11 @@ class WebUtils {
   /// Returns a stream broadcasting the fullscreen changes.
   static Stream<bool> get onFullscreenChange => const Stream.empty();
 
+  /// Indicates whether application is focused or not.
+  static bool get isFocused => false;
+
   /// Returns a stream broadcasting the browser's storage changes.
   static Stream<WebStorageEvent> get onStorageChange => const Stream.empty();
-
-  /// Returns a stream broadcasting the browser's page visibility.
-  static Stream<bool> get onVisibleChanged => const Stream.empty();
 
   /// Returns a stream broadcasting the browser's window focus changes.
   static Stream<bool> get onWindowFocus => const Stream.empty();